add Home-Component
