+++ public/index.html
@@ -7,6 +7,7 @@
     <link href="//fonts.googleapis.com/css?family=Raleway:400,300,600" rel="stylesheet" type="text/css">
     <link rel="stylesheet" href="css/normalize.css">
     <link rel="stylesheet" href="css/skeleton.css">
+    <link rel="stylesheet" href="css/styling.css">
     <script src="/hoodie/client.js"></script>
   </head>
   <body>
